better coloring

+++ utils.cpp
@@ -441,7 +441,7 @@ string bool2str(vector<bool> V){
 
 
 
-string color_coverage2str(const vector<uint32_t>& V){
+string color_coverage2str(const vector<uint16_t>& V){
 	string result;
 	for(uint64_t i(0);i<V.size();++i){
 		result+=":"+to_string(V[i]);
